Reworked creation of runnable worker to be created by the consumer actor. Mocked actors are now always single-threaded.
